+++ libraries/Stopwatch_RT/StopWatch.h
@@ -2,18 +2,17 @@
 //
 //    FILE: StopWatch.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.3.3
+// VERSION: 0.3.4
 // PURPOSE: Arduino Library implementing a stopwatch including seconds, milliseconds microseconds
 // HISTORY: See StopWatch.cpp
 //     URL: https://github.com/RobTillaart/StopWatch_RT
 //          http://playground.arduino.cc/Code/StopWatchClass
-//
 
 
 #include "Arduino.h"
 
 
-#define STOPWATCH_LIB_VERSION             (F("0.3.3"))
+#define STOPWATCH_LIB_VERSION             (F("0.3.4"))
 
 
 //  NOTE: adjust divider can fix timing inaccuracies (to some extend)
@@ -30,7 +29,7 @@ class StopWatch: public Printable
 {
 public:
   enum State      { RESET, RUNNING, STOPPED };
-  enum Resolution { MICROS, MILLIS, SECONDS, MINUTES};
+  enum Resolution { MICROS = 'u', MILLIS = 'm', SECONDS = 's', MINUTES = 'M' };
 
   explicit StopWatch(const enum Resolution resolution = MILLIS);
   void     start();
@@ -38,8 +37,12 @@ public:
   void     reset();
   //  elapsed() made const as printTo expect a const object.
   uint32_t elapsed() const;
+  char     getUnits();  //  returns u, s, m , M
 
   bool     isRunning();
+  bool     isStopped();
+  bool     isReset();
+
   enum     State state() { return _state; };
   void     setResolution(const enum Resolution resolution);
   enum     Resolution resolution() { return _resolution; };
@@ -52,8 +55,8 @@ private:
   enum State      _state;
   enum Resolution _resolution;
 
-  uint32_t _starttime;
-  uint32_t _stoptime;
+  uint32_t _startTime;
+  uint32_t _stopTime;
 
   uint32_t (*_gettime)(void);
   static uint32_t _micros()     { return micros(); };   //  wrapper for ESP proc.