+++ libraries/Stopwatch_RT/CHANGELOG.md
@@ -6,12 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.4] - 2023-01-11
+- update GitHub actions
+- update license
+- fix constructor - initialize all private variables.
+- add getUnits()
+- add isStopped() + isReset() convenience
+- update unit tests
+
+
 ## [0.3.3] - 2022-11-25
 - Add RP2040 support to build-CI.
 - Add CHANGELOG.md
 - declare value() obsolete
 - reordered .cpp file to match .h
-
 
 ## [0.3.2] - 2021-12-28  
 - update library.json